packages/shared/pyproject.toml
[project]
name = "shared"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "AnuragAmbekar741", email = "anuragambekar1997@gmail.com" },
]
requires-python = ">=3.12"
dependencies = [
    "aerich>=0.8.1",
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "sentry-sdk>=2.53.0",
    "tomli-w>=1.2.0",
    "tomlkit>=0.14.0",
    "tortoise-orm>=1.1.5",
]

[build-system]
requires = [
    "uv_build>=0.10.6,<0.11.0",
]
build-backend = "uv_build"

[tool.aerich]
tortoise_orm = "shared.db.connect.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."
